# swagger.yaml
openapi: 3.0.3
info:
  title: Jigi pay
  description: >-
    # A simple payment app api
  version: 1.0.0
  contact: { email: ammuftau74@gmail.com, whatapp: +2340935095173 }
servers:
  - url: https://jigipay.onrender.com/api/v1
    description: This is the original api URL
    schemes: [https, http]
  - url: http://localhost:8080/api/v1
    description: THis is my local Server
paths:
...
  /services/verify-phone:
    post:
      tags:
        - 👨‍💼 Services
      summary: Verify Phone
      description: >-
        Verify phone
      operationId: verifyPhone
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '09035095173'
            example:
              phone: '09035095173'
      responses:
        '200':
          description: Verify phone
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '85'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:59:47 GMT
            ETag:
              schema:
                type: string
                example: W/"55-CyGJCaxqcCXrzMloar9z7m95mvk"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
  /services/buy-airtime:
    post:
      tags:
        - 👨‍💼 Services
      summary: Buy data
      description: >-
        Buy data
      operationId: buyData
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '09035095173'
                network:
                  type: string
                  example: 1
                plan:
                  type: string
                  example: 500
            example:
              phone: '09035095173'
              network: 1
              plan: 500
      responses:
        '200':
          description: buy data
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '85'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:59:47 GMT
            ETag:
              schema:
                type: string
                example: W/"55-CyGJCaxqcCXrzMloar9z7m95mvk"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
  /services/buy-data:
    post:
      tags:
        - 👨‍💼 Services
      summary: Buy data
      description: >-
        Buy data
      operationId: buyData
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '09035095173'
                network:
                  type: string
                  example: 1
                plan:
                  type: string
                  example: 500
            example:
              phone: '09035095173'
              network: 1
              plan: 500
      responses:
        '200':
          description: buy data
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '85'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:59:47 GMT
            ETag:
              schema:
                type: string
                example: W/"55-CyGJCaxqcCXrzMloar9z7m95mvk"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
...
  /services/verify-bvn:
    post:
      tags:
        - 👨‍💼 Services
      summary: Verify BVN
      description: >-
        Verify bvn
      operationId: verifyBvn
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                  example: new
                lastname:
                  type: string
                  example: test
                phone:
                  type: string
                  example: 99098123
                bvn:
                  type: string
                  example: 878009123
            example:
              firstname: abdulsalam
              lastname: mohammed
              phone: '09012345789'
              bvn: '12134244'
      responses:
        '200':
          description: Verify bvn
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '85'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:59:47 GMT
            ETag:
              schema:
                type: string
                example: W/"55-CyGJCaxqcCXrzMloar9z7m95mvk"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
  /services/verify-nin:
    post:
      tags:
        - 👨‍💼 Services
      summary: Verify NIN
      description: >-
        Verify nin
      operationId: verifyNin
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                  example: new
                lastname:
                  type: string
                  example: test
                phone:
                  type: string
                  example: 99098123
                nin:
                  type: string
                  example: 878009123
            example:
              firstname: abdulsalam
              lastname: mohammed
              phone: '09005909513'
              nin: '37604587858'
      responses:
        '200':
          description: Verify nin
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '85'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:59:47 GMT
            ETag:
              schema:
                type: string
                example: W/"55-CyGJCaxqcCXrzMloar9z7m95mvk"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
...